pyproject.toml
[project]
name = "safai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.57.1",
    "google-genai>=1.24.0",
    "openai>=1.93.2",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "ruff>=0.12.2",
    "typer>=0.16.0",
]

[tool.ruff.lint]
    select = [
      "E",    
      "F",     
      "UP",   
      "RUF",  
      "N", 
    ]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
